Move some strategy functions to `tests::setup`
Create a new module for the strategy functions that are only used
internally.
