+++ zebra-state/src/service/check/tests/utxo.rs
@@ -14,10 +14,8 @@ use zebra_chain::{
 
 use crate::{
     arbitrary::Prepare,
-    service::{
-        arbitrary::{new_state_with_mainnet_genesis, transaction_v4_from_coinbase},
-        StateService,
-    },
+    service::StateService,
+    tests::setup::{new_state_with_mainnet_genesis, transaction_v4_from_coinbase},
     FinalizedBlock,
     ValidateContextError::{
         DuplicateTransparentSpend, EarlyTransparentSpend, MissingTransparentOutput,
